pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.5",
    "openai>=1.77.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "poppler-utils>=0.1.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "requests>=2.32.3",
]
